Clean version first try.
I want to clean the VRTK version to be able to have a PC mode and based on this version to be able to add other VR libraries in the future without having to depend on obsolete libraries. VR changes so much that it's better this way. This is only preliminary and may have many flaws.